% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{barycenter_unbalanced_stabilized}
\alias{barycenter_unbalanced_stabilized}
\title{unbalanced barycenter}
\usage{
barycenter_unbalanced_stabilized(
  A,
  M,
  weights,
  reg = 0.1,
  reg_m = 1,
  numItermax = 1000L,
  stopThr = 1e-08,
  verbose = FALSE
)
}
\arguments{
\item{A}{matrix(dim, n). n training distributions a_i of dimension dim.}

\item{M}{matrix(dim, dim). ground metric matrix for OT.}

\item{reg}{Entropy regularization term > 0.}

\item{reg_m}{Marginal relaxation term > 0.}

\item{numItermax}{int. Max number of iterations.}

\item{stopThr}{float. Stop threshol on error (> 0).}

\item{verbose}{bool. Print information along iterations.}

\item{weighs}{vector(n). Weight of each distribution (barycentric coodinates).}
}
\value{
vector(dim). Unbalanced Wasserstein barycenter
}
\description{
Compute the entropic unbalanced wasserstein barycenter of A.
}
\examples{
a=dnorm(1:100,20,10)
a=a/sum(a)
b=dnorm(1:100, mean=60, sd=2)
b=b/sum(b)*2
dab = matrix(c(a, b), 100, 2)
distance= dists(1:100,1:100)
distance = distance/max(distance)
res = barycenter_unbalanced_stabilized(dab, distance, reg = 1, reg_m = 1, weights = c(0.5, 0.5))
p = data.frame(ind = 1:100, a = a, b = b, c = res)
ggplot(data = p)+geom_line(aes(x = ind, y = a), color = "blue")+
geom_line(aes(x = ind, y = b), color = "red")+
geom_line(aes(x = ind, y = res), color = "green")

}
\references{
[1] Benamou, J. D., Carlier, G., Cuturi, M., Nenna, L., & Peyre, G.(2015). Iterative Bregman projections for regularized transportation problems. SIAM Journal on Scientific Computing, 37(2), A1111-A1138.
[2] Chizat, L., Peyre, G., Schmitzer, B., & Vialard, F. X. (2016). Scaling algorithms for unbalanced transport problems. arXiv preprin arXiv:1607.05816.
}
