% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{dist_mat}
\alias{dist_mat}
\title{distance between two samples(dimension>1)}
\usage{
dist_mat(a, b, k = 1L, square = 0L)
}
\arguments{
\item{a}{mat a(n,p) n:the number of samples,p:the dimensionnality}

\item{b}{mat b(n,p) n:the number of samples,p:the dimensionnality}

\item{k}{the order of the norm of the distance (default k=1)}
}
\value{
the k-norm of the distance between a and b
}
\description{
Compute the distance matrix between two samples (dimensions>1)
}
\examples{
a = cbind(1:10,1:10)
dist_mat(a,a)
}
