% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{dists}
\alias{dists}
\title{distance between two samples(dimension=1)}
\usage{
dists(a, b, k = 1L, dimention = 1L)
}
\arguments{
\item{a}{sample a}

\item{b}{sample b}

\item{k}{the order of the norm (default k=1)}
}
\value{
the kth power of the k-norm of the distance between a and b
}
\description{
Compute the distance matrix between samples of one dimention.
}
\examples{
distance=dists(1:100,1:100)
}
