% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{pat_uot_was}
\alias{pat_uot_was}
\title{computer the wasserstein distance  of partial optimal transport}
\usage{
pat_uot_was(a, b, dab, reg)
}
\arguments{
\item{a}{the the vector of source distribution}

\item{b}{the vector of target dirtribution}

\item{reg}{entropic regularization parameter}

\item{mat}{the cost matrix}
}
\description{
computer the wasserstein distance  of partial optimal transport
}
\examples{
a<-c(0.4,0.8)
b<-c(0.6,3)
m1<-c(0,1)
m2<-c(1,0)
m<-cbind(m1,m2)
pat_uot_was(a,b,m,1)
}
