% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot2.R
\name{plot1D}
\alias{plot1D}
\title{Plot 1D distribution}
\usage{
plot1D(a, b, M, title = "")
}
\arguments{
\item{a}{Source distribution}

\item{b}{Target distribution}

\item{M}{Matrix to plot}

\item{title}{The title of the plot}
}
\description{
Plot matrix M with the source and target 1D distribution
Creates a subplot with the source distribution a on the left and target distribution b on the tot. The matrix M is shown in between.
}
\examples{
## SMALL EXAMPLE
n=10
a=dnorm(1:100,20,10)
a=a/sum(a)
b=dnorm(1:100, mean=60, sd=2)
b=b/sum(b)*2
distance= dists(1:100,1:100)
distance= diatance/max(distance)
plan=sinkhorn_knopp_unbalanced(a,b,distance,1,1,1000,1e-6,TRUE,FALSE)
plot1D(a, b, plan, title = "UOT example")
}
