% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot2.R
\name{plot2D}
\alias{plot2D}
\title{Plot 2D distribution}
\usage{
plot2D(xs, xt, G, thr = 1e-08, title = "")
}
\arguments{
\item{thr}{threshold above which the line is drawn}

\item{title}{The title of the plot}

\item{a}{an \eqn{(M\times 2)} matrix of Source samples positions}

\item{b}{an \eqn{(N\times 2)} matrix of Target samples positions}

\item{M}{OT matrix}
}
\description{
Plot matrix M in 2D with lines using alpha values.
}
\examples{
## SMALL EXAMPLE
n = 30
mu_s = c(0, 0)
cov_s = matrix(c(1, 0, 0, 1), 2, 2)
mu_t = c(4, 4)
cov_t = matrix(c(1, -0.5, -0.5, 1), 2, 2)
library(MASS)
library(transport)
xs = mvrnorm(n, mu_s, cov_s)
xt = mvrnorm(n, mu_t, cov_t)
a = rep(1, n)/n
b = rep(1, n)/n
M = dist_mat(xs, xt)
M = M/max(M)
plan = transport(a, b, M)
p = rep(0, n^2)
p[(plan$to-1)*n + plan$from] = plan$mass
p = matrix(p, n, n)
plot2D(xs, xt, p, title = "2DOT example")
}
