% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{sinkhorn_knopp_unbalanced}
\alias{sinkhorn_knopp_unbalanced}
\title{computer the plan of regularized unbalanced optimal transport}
\usage{
sinkhorn_knopp_unbalanced(
  a,
  b,
  dab,
  reg,
  reg_m,
  numItermax = 100L,
  stopThr = 1e-06
)
}
\arguments{
\item{a}{vec a is the the histogram of source distribution(dim_a=m)}

\item{b}{vec b is the histogram of target dirtribution (dim_b=n)}

\item{dab}{matrix dab is the cost matrix between a and b}

\item{reg}{entropic regularization parameter}

\item{reg_m}{marginal relaxation term > 0}

\item{stopThr}{ending condition}

\item{numItemax}{the max iteration times}
}
\value{
a matrix that shape:m
}
\description{
Solve the entropic regularization unbalanced optimal transport
}
\examples{
a<-c(0.4,0.8)
b<-c(0.6,3)
m1<-c(0,1)
m2<-c(1,0)
m<-cbind(m1,m2)
sinkhorn_knopp_unbalanced(a,b,m,1,1)
}
