% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{sinkhorn_stabilized_was}
\alias{sinkhorn_stabilized_was}
\title{sinkhorn unbalanced OT distance}
\usage{
sinkhorn_stabilized_was(a, b, dab, reg, reg_m)
}
\arguments{
\item{a}{vector(dim_a). Unnormalized histogram of dimension dim_a.}

\item{b}{vector(dim_b). Unnormalized histogram of dimension dim_b.}

\item{dab}{matrix(dim_a, dim_b). loss matrix.}

\item{reg}{float. Entropy regularization term > 0.}

\item{reg_m}{float. Marginal relaxation term > 0.}
}
\value{
the OT distance between `a` and 'b`
}
\description{
Solve the entropic regularization unbalanced optimal transport problem and return the loss.
}
\examples{
a<-c(0.4,0.8)
b<-c(0.6,3)
m1<-c(0,1)
m2<-c(1,0)
m<-cbind(m1,m2)
sinkhorn_stabilized_was(a,b,m,1,1)

}
\references{
[1] Chizat, L., Peyre, G., Schmitzer, B., & Vialard, F. X. (2016). Scaling algorithms for unbalanced transport problems. arXiv preprint arXiv:1607.05816.
[2] Frogner C., Zhang C., Mobahi H., Araya-Polo M., Poggio T. Learning with a Wasserstein Loss,  Advances in Neural Information Processing Systems (NIPS) 2015.
}
